+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Sun Sep 27 16:02:30 BST 2020
+gradle.version=6.5
