+++ connector/src/main/java/org/geysermc/connector/network/session/GeyserSession.java
@@ -143,7 +143,7 @@ public class GeyserSession implements CommandSender {
 
     public void setSneaking(boolean sneaking) {
         this.sneaking = sneaking;
-        updatePlayerBoundingBox();
+        collisionManager.updatePlayerBoundingBox();
     }
 
     @Setter
@@ -152,8 +152,9 @@ public class GeyserSession implements CommandSender {
     @Setter
     private boolean jumping;
 
-    @Getter
-    private BoundingBox playerBoundingBox;
+    // TODO: Remove
+    // @Getter
+    // private BoundingBox playerBoundingBox;
 
     @Setter
     private int breakingBlock;
@@ -275,10 +276,9 @@ public class GeyserSession implements CommandSender {
     private MinecraftProtocol protocol;
 
     /**
-     * Additional space where blocks are checked, which is helpful for fixing NoCheatPlus's Passable check.
-     * This check doesn't allow players right up against the block, so they must be pushed slightly away.
+     * Stores session collision
      */
-    public static final double COLLISION_TOLERANCE = 0.00001;
+    private CollisionManager collisionManager;
 
     public GeyserSession(GeyserConnector connector, BedrockServerSession bedrockServerSession) {
         this.connector = connector;
@@ -290,6 +290,8 @@ public class GeyserSession implements CommandSender {
         this.worldCache = new WorldCache(this);
         this.windowCache = new WindowCache(this);
 
+        this.collisionManager = new CollisionManager(this);
+
         this.playerEntity = new PlayerEntity(new GameProfile(UUID.randomUUID(), "unknown"), 1, 1, Vector3f.ZERO, Vector3f.ZERO, Vector3f.ZERO, this);
         this.inventory = new PlayerInventory();
 
@@ -299,8 +301,6 @@ public class GeyserSession implements CommandSender {
         this.loggedIn = false;
 
         this.inventoryCache.getInventories().put(0, inventory);
-
-        this.playerBoundingBox = new BoundingBox(0, 0, 0, 0.6, 1.8, 0.6);
 
         bedrockServerSession.addDisconnectHandler(disconnectReason -> {
             connector.getLogger().info(LanguageUtils.getLocaleStringLog("geyser.network.disconnect", bedrockServerSession.getAddress().getAddress(), disconnectReason));
@@ -808,104 +808,5 @@ public class GeyserSession implements CommandSender {
 
         adventureSettingsPacket.getSettings().addAll(flags);
         sendUpstreamPacket(adventureSettingsPacket);
-    }
-
-    /**
-     * Updates the stored bounding box
-     * @param position The new position of the player
-     */
-    public void updatePlayerBoundingBox(Vector3f position) {
-        updatePlayerBoundingBox(Vector3d.from(position.getX(), position.getY(), position.getZ()));
-    }
-
-    /**
-     * Updates the stored bounding box
-     * @param position The new position of the player
-     */
-    public void updatePlayerBoundingBox(Vector3d position) {
-        updatePlayerBoundingBox();
-
-        playerBoundingBox.setMiddleX(position.getX());
-        playerBoundingBox.setMiddleY(position.getY() + (playerBoundingBox.getSizeY() / 2));
-        playerBoundingBox.setMiddleZ(position.getZ());
-    }
-
-    /**
-     * Updates the stored bounding box without passing a position, which currently just changes the height depending on if the player is sneaking.
-     */
-    public void updatePlayerBoundingBox() {
-        if (playerBoundingBox == null) {
-            Vector3f playerPosition;
-            if (playerEntity == null) {
-                // Temporary position to prevent NullPointerException
-                playerPosition = Vector3f.ZERO;
-            } else {
-                playerPosition = playerEntity.getPosition();
-            }
-            playerBoundingBox = new BoundingBox(playerPosition.getX(), playerPosition.getY() + 0.9, playerPosition.getZ(), 0.6, 1.8, 0.6);
-        } else {
-            // According to the Minecraft Wiki, when sneaking:
-            // - In Bedrock Edition, the height becomes 1.65 blocks, allowing movement through spaces as small as 1.75 (2 - 1⁄4) blocks high.
-            // - In Java Edition, the height becomes 1.5 blocks.
-            if (sneaking) {
-                playerBoundingBox.setSizeY(1.5);
-            } else {
-                playerBoundingBox.setSizeY(1.8);
-            }
-        }
-    }
-
-    public List<Vector3i> getPlayerCollidableBlocks() {
-        List<Vector3i> blocks = new ArrayList<>();
-
-        Vector3d position = Vector3d.from(playerBoundingBox.getMiddleX(),
-                playerBoundingBox.getMiddleY() - (playerBoundingBox.getSizeY() / 2),
-                playerBoundingBox.getMiddleZ());
-
-        // Loop through all blocks that could collide with the player
-        int minCollisionX = (int) Math.floor(position.getX() - ((playerBoundingBox.getSizeX() / 2) + COLLISION_TOLERANCE));
-        int maxCollisionX = (int) Math.floor(position.getX() + (playerBoundingBox.getSizeX() / 2) + COLLISION_TOLERANCE);
-
-        // Y extends 0.5 blocks down because of fence hitboxes
-        int minCollisionY = (int) Math.floor(position.getY() - 0.5);
-
-        int maxCollisionY = (int) Math.floor(position.getY() + playerBoundingBox.getSizeY());
-
-        int minCollisionZ = (int) Math.floor(position.getZ() - ((playerBoundingBox.getSizeZ() / 2) + COLLISION_TOLERANCE));
-        int maxCollisionZ = (int) Math.floor(position.getZ() + (playerBoundingBox.getSizeZ() / 2) + COLLISION_TOLERANCE);
-
-        for (int y = minCollisionY; y < maxCollisionY + 1; y++) {
-            for (int x = minCollisionX; x < maxCollisionX + 1; x++) {
-                for (int z = minCollisionZ; z < maxCollisionZ + 1; z++) {
-                    blocks.add(Vector3i.from(x, y, z));
-                }
-            }
-        }
-
-        return blocks;
-    }
-
-    public void correctPlayerPosition() {
-        List<Vector3i> collidableBlocks = getPlayerCollidableBlocks();
-
-        // Used when correction code needs to be run before the main correction
-        for (Vector3i blockPos : collidableBlocks) {
-            BlockCollision blockCollision = CollisionTranslator.getCollisionAt(
-                    blockPos.getX(), blockPos.getY(), blockPos.getZ(), this
-            );
-            if (blockCollision != null) {
-                blockCollision.beforeCorrectPosition(playerBoundingBox);
-            }
-        }
-
-        // Main correction code
-        for (Vector3i blockPos : collidableBlocks) {
-            BlockCollision blockCollision = CollisionTranslator.getCollisionAt(
-                    blockPos.getX(), blockPos.getY(), blockPos.getZ(), this
-            );
-            if (blockCollision != null) {
-                blockCollision.correctPosition(playerBoundingBox);
-            }
-        }
     }
 }
